refactor: move extensions to right location
